refactor(import): split CsvUploadStep into TimezonePicker and CsvDropZone
CsvUploadStep was carrying a Select, a drop zone, hidden input, helper
text, and the action button all in one component. Pull the timezone Select
into TimezonePicker and the drop zone (icon, file name, row count, hidden
input, format helper) into CsvDropZone. The parent now reads as the actual
flow: pick a timezone, drop a file, hit analyse.
