+++ src/components/Admin/ScheduleImport/TimezonePicker.tsx
@@ -1,0 +1,45 @@
+import { Label } from "@/components/ui/label";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+
+const TIMEZONES = [
+  { value: "Europe/Lisbon", label: "Lisbon (WET/WEST)" },
+  { value: "Europe/London", label: "London (GMT/BST)" },
+  { value: "Europe/Berlin", label: "Berlin (CET/CEST)" },
+  { value: "America/New_York", label: "New York (EST/EDT)" },
+  { value: "America/Los_Angeles", label: "Los Angeles (PST/PDT)" },
+  { value: "UTC", label: "UTC" },
+];
+
+type Props = {
+  value: string;
+  onChange: (value: string) => void;
+};
+
+export function TimezonePicker({ value, onChange }: Props) {
+  return (
+    <div className="space-y-2">
+      <Label>Timezone</Label>
+      <Select value={value} onValueChange={onChange}>
+        <SelectTrigger className="w-64">
+          <SelectValue />
+        </SelectTrigger>
+        <SelectContent>
+          {TIMEZONES.map((tz) => (
+            <SelectItem key={tz.value} value={tz.value}>
+              {tz.label}
+            </SelectItem>
+          ))}
+        </SelectContent>
+      </Select>
+      <p className="text-sm text-muted-foreground">
+        All times in the CSV are interpreted as local festival time.
+      </p>
+    </div>
+  );
+}
